Add command executer class
